Attempt to use shared Koza mock writer classes in unit tests for all current ingests
