+++ tests/unit/ingests/ctd/test_ctd.py
@@ -1,39 +1,21 @@
-from typing import Iterable
-
 import pytest
 from biolink_model.datamodel.pydanticmodel_v2 import (
     ChemicalToDiseaseOrPhenotypicFeatureAssociation,
     ChemicalEntity,
     Disease, RetrievalSource,
 )
-from koza.io.writer.writer import KozaWriter
-from koza.runner import KozaRunner, KozaTransformHooks
-from src.translator_ingest.ingests.ctd.ctd import transform_record_chemical_to_disease as ctd_transform
 
+from koza.runner import KozaRunner, KozaTransformHooks
+from translator_ingest.ingests.ctd.ctd import transform_record_chemical_to_disease as ctd_transform
+
+from tests.unit.ingests import MockKozaWriter
 
 BIOLINK_TREATS_OR_APPLIED_OR_STUDIED_TO_TREAT = "biolink:treats_or_applied_or_studied_to_treat"
 
 
-class MockWriter(KozaWriter):
-    def __init__(self):
-        self.items = []
-
-    def write(self, entities):
-        self.items += entities
-
-    def write_nodes(self, nodes: Iterable):
-        self.items += nodes
-
-    def write_edges(self, edges: Iterable):
-        self.items += edges
-
-    def finalize(self):
-        pass
-
-
 @pytest.fixture
 def therapeutic_output():
-    writer = MockWriter()
+    writer = MockKozaWriter()
     record = {
         "ChemicalName": "10,11-dihydro-10-hydroxycarbamazepine",
         "ChemicalID": "C039775",
